stock-ticker-bot.py
import os
import tweepy

from google.cloud import storage


def cashtag_scrape():
    consumer_key = os.environ['TWITTER_CONSUMER_KEY']
    consumer_secret = os.environ['TWITTER_CONSUMER_SECRET']
    access_token = os.environ['TWITTER_ACCESS_TOKEN']
    access_token_secret = os.environ['TWITTER_ACCESS_TOKEN_SECRET']
    

def main():
    scrape()


if __name__ == "__main__":
    main()
